Move all overlay creation function to a separate file
